Backend/casting/requirements.txt
Flask
Flask-Mail
Flask-pymongo
Flask-Cors
python-dotenv
Flask-JWT-Extended